+++ test/lib/audit-fs/index.js
@@ -11,8 +11,10 @@ chai.use(require('sinon-chai'));
 
 var auditFs = require('../../..');
 var AuditFs = auditFs.AuditFs;
+var rules = auditFs.rules;
 
 require('sinon-doublist')(sinon, 'mocha');
+require('sinon-doublist-fs')('mocha');
 
 describe('AuditFs', function() {
   'use strict';
@@ -24,16 +26,17 @@ describe('AuditFs', function() {
     this.testRes = {iAmA: 'test result'};
     this.dir = 'dir';
     this.file = 'file.ext';
-    this.as = new auditFs.create();
+    this.afs = new auditFs.create();
     this.resOK = {code: 0};
     this.hasDirSpy = this.spy(auditFs.rules, 'hasDir');
     this.hasFileStub = this.stub(auditFs.rules, 'hasFile');
+    this.now = Date.now();
   });
 
   describe('constructor', function() {
     it('should use cwd as default target dir', function() {
-      this.as = new auditFs.create();
-      this.as.get('dir').should.equal(process.cwd());
+      this.afs = new auditFs.create();
+      this.afs.get('dir').should.equal(process.cwd());
     });
   });
 
@@ -41,7 +44,7 @@ describe('AuditFs', function() {
     it('should include generated rules', function() {
       var self = this;
       Object.keys(auditFs.rules).forEach(function(name) {
-        self.as[name].should.be.a('function');
+        self.afs[name].should.be.a('function');
       });
     });
   });
@@ -61,62 +64,62 @@ describe('AuditFs', function() {
   describe('generated test wrapper', function() {
     it('should skip current test if a past test failed', function() {
       this.hasFileStub.withArgs(this.file).returns(false);
-      this.as.hasFile(this.file).hasDir(this.dir).pass();
+      this.afs.hasFile(this.file).hasDir(this.dir).pass();
       this.hasFileStub.should.have.been.called;
       this.hasDirSpy.should.not.have.been.calledWith(this.dir);
     });
 
     it('should pass-through all args', function() {
       var _Stub = this.stub(auditFs.rules, '_');
-      this.as._('foo', 'bar', 'baz').pass();
+      this.afs._('foo', 'bar', 'baz').pass();
       _Stub.should.have.been.calledWithExactly('foo', 'bar', 'baz');
     });
 
     it('should save results', function() {
       this.hasFileStub.withArgs(this.file).returns(this.testRes);
-      this.as.hasFile(this.file).pass();
-      this.as.results.length.should.equal(1);
-      this.as.results[0].name.should.equal('hasFile');
-      this.as.results[0].res.should.deep.equal(this.testRes);
-      this.as.results[0].args.should.deep.equal([this.file]);
+      this.afs.hasFile(this.file).pass();
+      this.afs.results.length.should.equal(1);
+      this.afs.results[0].name.should.equal('hasFile');
+      this.afs.results[0].res.should.deep.equal(this.testRes);
+      this.afs.results[0].args.should.deep.equal([this.file]);
     });
 
     it('should not change match status on pass', function() {
       this.hasFileStub.withArgs(this.file).returns(true);
-      this.as.match.should.equal(true);
-      this.as.hasFile(this.file).pass();
-      this.as.match.should.equal(true);
+      this.afs.match.should.equal(true);
+      this.afs.hasFile(this.file).pass();
+      this.afs.match.should.equal(true);
     });
 
     it('should correctly update match status on miss', function() {
       this.hasFileStub.withArgs(this.file).returns(false);
-      this.as.match.should.equal(true);
-      this.as.hasFile(this.file).pass();
-      this.as.match.should.equal(false);
+      this.afs.match.should.equal(true);
+      this.afs.hasFile(this.file).pass();
+      this.afs.match.should.equal(false);
     });
   });
 
   describe('#pass', function() {
     it('should begin commands from target dir', function() {
-      this.as.set('dir', this.dir);
-      var stub = this.stub(this.as.shelljs, '_');
-      this.as.pass();
+      this.afs.set('dir', this.dir);
+      var stub = this.stub(this.afs.shelljs, '_');
+      this.afs.pass();
       stub.should.have.been.calledWith('cd', this.dir);
     });
 
     it('should return match result', function() {
-      this.as.pass().should.equal(true);
+      this.afs.pass().should.equal(true);
 
       this.hasFileStub.withArgs(this.file).returns(false);
-      this.as.hasFile(this.file).pass().should.equal(false);
+      this.afs.hasFile(this.file).pass().should.equal(false);
     });
   });
 
   describe('#last', function() {
     it('should return last test result', function() {
       this.hasFileStub.withArgs(this.file).returns(false);
-      this.as.hasDir('').hasFile(this.file).pass();
-      this.as.last().should.deep.equal({
+      this.afs.hasDir('').hasFile(this.file).pass();
+      this.afs.last().should.deep.equal({
         name: 'hasFile', args: [this.file], res: false
       });
     });
@@ -124,20 +127,20 @@ describe('AuditFs', function() {
 
   describe('test', function() {
     beforeEach(function() {
-      this._Stub = this.stub(this.as.shelljs, '_');
+      this._Stub = this.stub(this.afs.shelljs, '_');
     });
 
     describe('#_', function() {
       it('should return shelljs pass-through result', function() {
-        this.as._('test', '-L', this.file).pass();
+        this.afs._('test', '-L', this.file).pass();
         this._Stub.should.have.been.calledWithExactly('test', '-L', this.file);
       });
     });
 
     describe('#__', function() {
       it('should return outer-shelljs pass-through result', function() {
-        var stub = this.stub(this.as.shelljs, 'grep');
-        this.as.__('grep', 'needle', '/path/to/haystack').pass();
+        var stub = this.stub(this.afs.shelljs, 'grep');
+        this.afs.__('grep', 'needle', '/path/to/haystack').pass();
         stub.should.have.been.calledWithExactly('needle', '/path/to/haystack');
       });
     });
@@ -145,21 +148,21 @@ describe('AuditFs', function() {
     describe('#assert', function() {
       it('should receive OuterShelljs instance', function() {
         var cb = this.stub();
-        this.as.assert('some expectation', cb).pass();
-        cb.should.have.been.calledWithExactly(this.as.shelljs);
+        this.afs.assert('some expectation', cb).pass();
+        cb.should.have.been.calledWithExactly(this.afs.shelljs);
       });
 
       it('should return custom function pass', function() {
         var cb = this.stub();
         cb.returns(false);
-        this.as.assert('some expectation', cb).pass().should.equal(false);
+        this.afs.assert('some expectation', cb).pass().should.equal(false);
         cb.should.have.been.called;
       });
 
       it('should return custom function fail', function() {
         var cb = this.stub();
         cb.returns(true);
-        this.as.assert('some expectation', cb).pass().should.equal(true);
+        this.afs.assert('some expectation', cb).pass().should.equal(true);
         cb.should.have.been.called;
       });
     });
@@ -168,21 +171,21 @@ describe('AuditFs', function() {
       describe('function', function() {
         it('should receive OuterShelljs instance', function() {
           var cb = this.stub();
-          this.as.refute('some expectation', cb).pass();
-          cb.should.have.been.calledWithExactly(this.as.shelljs);
+          this.afs.refute('some expectation', cb).pass();
+          cb.should.have.been.calledWithExactly(this.afs.shelljs);
         });
 
         it('should return custom function pass', function() {
           var cb = this.stub();
           cb.returns(true);
-          this.as.refute('some expectation', cb).pass().should.equal(false);
+          this.afs.refute('some expectation', cb).pass().should.equal(false);
           cb.should.have.been.called;
         });
 
         it('should return custom function fail', function() {
           var cb = this.stub();
           cb.returns(false);
-          this.as.refute('some expectation', cb).pass().should.equal(true);
+          this.afs.refute('some expectation', cb).pass().should.equal(true);
           cb.should.have.been.called;
         });
 
@@ -190,9 +193,9 @@ describe('AuditFs', function() {
           var self = this;
           Object.keys(auditFs.rules).forEach(function(name) {
             if (name === 'assert' || name === 'refute') {
-              should.not.exist(self.as.refute[name]);
+              should.not.exist(self.afs.refute[name]);
             } else {
-              self.as.refute[name].should.be.a('function');
+              self.afs.refute[name].should.be.a('function');
             }
           });
         });
@@ -203,7 +206,7 @@ describe('AuditFs', function() {
           this.hasFileStub.restore();
           var exists = true;
           this._Stub.returns(exists);
-          this.as.refute.hasFile(this.file).pass().should.equal(!exists);
+          this.afs.refute.hasFile(this.file).pass().should.equal(!exists);
           this._Stub.should.have.been.calledWithExactly(
             'test', '-f', process.cwd() + '/' + this.file
           );
@@ -223,53 +226,53 @@ describe('AuditFs', function() {
 
       describe('#grep variant', function() {
         it('should pass-through args to #exec', function() {
-          this.as.grep(this.textPat, this.filePat).pass();
+          this.afs.grep(this.textPat, this.filePat).pass();
           this._Stub.should.have.been.calledWith(
             'exec', ['grep', '-l', this.textPatFinal, this.filePat].join(' ')
           );
         });
 
         it('should return true on match', function() {
-          this.as.grep(this.textPat, this.filePat).pass().should.equal(true);
+          this.afs.grep(this.textPat, this.filePat).pass().should.equal(true);
         });
 
         it('should return false on match', function() {
           this.res.code = 1;
           this.res.output = '';
-          this.as.grep(this.textPat, this.filePat).pass().should.equal(false);
+          this.afs.grep(this.textPat, this.filePat).pass().should.equal(false);
         });
 
         it('should return false on error', function() {
           this.res.code = 2;
-          this.as.grep(this.textPat, this.filePat).pass().should.equal(false);
+          this.afs.grep(this.textPat, this.filePat).pass().should.equal(false);
         });
       });
 
       describe('#grepv variant', function() {
         it('should pass-through args to #exec', function() {
-          this.as.grepv(this.textPat, this.filePat).pass();
+          this.afs.grepv(this.textPat, this.filePat).pass();
           this._Stub.should.have.been.calledWith(
             'exec', ['grep', '-vl', this.textPatFinal, this.filePat].join(' ')
           );
         });
 
         it('should return true on match', function() {
-          this.as.grepv(this.textPat, this.filePat).pass().should.equal(true);
+          this.afs.grepv(this.textPat, this.filePat).pass().should.equal(true);
         });
 
         it('should return false on no match', function() {
           this.res.code = 1;
           this.res.output = '';
-          this.as.grep(this.textPat, this.filePat).pass().should.equal(false);
+          this.afs.grep(this.textPat, this.filePat).pass().should.equal(false);
         });
 
         it('should return false on error', function() {
           this.res.code = 2;
-          this.as.grep(this.textPat, this.filePat).pass().should.equal(false);
+          this.afs.grep(this.textPat, this.filePat).pass().should.equal(false);
         });
 
         it('should merge flags', function() {
-          this.as.grepv('-r', this.textPat, this.filePat).pass().should.equal(true);
+          this.afs.grepv('-r', this.textPat, this.filePat).pass().should.equal(true);
           this._Stub.should.have.been.calledWith(
             'exec', ['grep', '-rvl', this.textPatFinal, this.filePat].join(' ')
           );
@@ -277,10 +280,18 @@ describe('AuditFs', function() {
       });
     });
 
+    describe('#exists', function() {
+      it('should return shelljs pass-through result', function() {
+        this._Stub.returns(true);
+        this.afs.exists('').pass().should.equal(true);
+        this._Stub.should.have.been.calledWithExactly('test', '-e', process.cwd() + '/');
+      });
+    });
+
     describe('#hasDir', function() {
       it('should return shelljs pass-through result', function() {
         this._Stub.returns(true);
-        this.as.hasDir('').pass().should.equal(true);
+        this.afs.hasDir('').pass().should.equal(true);
         this._Stub.should.have.been.calledWithExactly('test', '-d', process.cwd() + '/');
       });
     });
@@ -289,10 +300,143 @@ describe('AuditFs', function() {
       it('should return shelljs pass-through result', function() {
         this.hasFileStub.restore();
         this._Stub.returns(true);
-        this.as.hasFile(this.file).pass().should.equal(true);
+        this.afs.hasFile(this.file).pass().should.equal(true);
         this._Stub.should.have.been.calledWithExactly(
           'test', '-f', process.cwd() + '/' + this.file
         );
+      });
+    });
+
+    describe('#getFileSize', function() {
+      it('should calculate file size', function() {
+        this.stubFile('/f').stat('size', 1234).make();
+        AuditFs.getFileSize('/f').should.equal(1234);
+      });
+
+      it('should calculate directory size recursively', function() {
+        this.stubFile('/d').readdir([
+          this.stubFile('/d/f0').stat('size', 1),
+          this.stubFile('/d/sub0').readdir([
+            this.stubFile('/d/sub0/f1').stat('size', 2),
+            this.stubFile('/d/sub0/f2').stat('size', 3),
+            this.stubFile('/d/sub0/sub1').readdir([
+              this.stubFile('/d/sub0/sub1/f3').stat('size', 4)
+            ]).stat('size', 4096)
+          ]).stat('size', 4096),
+          this.stubFile('/d/sub2').readdir([
+            this.stubFile('/d/sub2/f4').stat('size', 5)
+          ]).stat('size', 4096)
+        ]).stat('size', 4096).make();
+        AuditFs.getFileSize('/d').should.equal(16399);
+      });
+    });
+
+    describe('#getFileCount', function() {
+      it('should calculate directory file count non-recursively', function() {
+        this.stubFile('/d').readdir([
+          this.stubFile('/d/f0'),
+          this.stubFile('/d/sub0').readdir(['f3', 'f4', 'sub1']),
+          this.stubFile('/d/f1'),
+          this.stubFile('/d/f2')
+        ]).make();
+        AuditFs.getFileCount('/d').should.equal(3);
+      });
+    });
+
+    describe('#minSize', function() {
+      it('should fail if min unmet', function() {
+        this.stubFile('/f0').stat('size', 3).make();
+        rules.minSize({filename: '/f0', size: 4}).should.equal(false);
+      });
+
+      it('should pass if min is met exactly', function() {
+        this.stubFile('/f0').stat('size', 4).make();
+        rules.minSize({filename: '/f0', size: 4}).should.equal(true);
+      });
+
+      it('should pass if min is exceeded', function() {
+        this.stubFile('/f0').stat('size', 5).make();
+        rules.minSize({filename: '/f0', size: 4}).should.equal(true);
+      });
+    });
+
+    describe('#maxSize', function() {
+      it('should fail if max exceeded', function() {
+        this.stubFile('/f0').stat('size', 4).make();
+        rules.maxSize({filename: '/f0', size: 3}).should.equal(false);
+      });
+
+      it('should pass if max is met exactly', function() {
+        this.stubFile('/f0').stat('size', 4).make();
+        rules.maxSize({filename: '/f0', size: 4}).should.equal(true);
+      });
+
+      it('should pass if max is unmet', function() {
+        this.stubFile('/f0').stat('size', 3).make();
+        rules.maxSize({filename: '/f0', size: 4}).should.equal(true);
+      });
+    });
+
+    describe('#minCount', function() {
+      beforeEach(function() {
+        this.stubFile('/f').readdir([
+          this.stubFile('/f/a'),
+          this.stubFile('/f/b'),
+          this.stubFile('/f/c')
+        ]).make();
+      });
+
+      it('should pass if min is met exactly', function() {
+        rules.minCount({filename: '/f', count: 3}).should.equal(true);
+      });
+
+      it('should pass if min is exceeded', function() {
+        rules.minCount({filename: '/f', count: 2}).should.equal(true);
+      });
+    });
+
+    describe('#maxCount', function() {
+      beforeEach(function() {
+        this.stubFile('/f').readdir(['a', 'b', 'c']).make();
+        this.stubFile('/f/a').make();
+        this.stubFile('/f/b').make();
+        this.stubFile('/f/c').make();
+      });
+
+      it('should fail if max exceeded', function() {
+        rules.maxCount({filename: '/f', count: 2}).should.equal(false);
+      });
+
+      it('should pass if max is met exactly', function() {
+        rules.maxCount({filename: '/f', count: 3}).should.equal(true);
+      });
+
+      it('should pass if max is unmet', function() {
+        rules.maxCount({filename: '/f', count: 3}).should.equal(true);
+      });
+    });
+
+    describe('#created', function() {
+      it('should fail if file was not created recently', function() {
+        this.stubFile('/f').stat('ctime', new Date(this.now - 10)).make();
+        rules.created({filename: '/f', max: 9}).should.equal(false);
+      });
+
+      it('should pass if file was created recently', function() {
+        this.stubFile('/f').stat('ctime', new Date(this.now - 1)).make();
+        rules.created({filename: '/f', max: 9}).should.equal(true);
+      });
+    });
+
+    describe('#modified', function() {
+      it('should fail if file was not updated recently', function() {
+        this.stubFile('/f').stat('mtime', new Date(this.now - 10)).make();
+        rules.modified({filename: '/f', max: 9}).should.equal(false);
+      });
+
+      it('should pass if file was updated recently', function() {
+        this.stubFile('/f').stat('mtime', new Date(this.now - 1)).make();
+        rules.modified({filename: '/f', max: 9}).should.equal(true);
       });
     });
   });